requirements.txt
einops==0.6.0
numpy==1.23.4
Pillow==9.3.0
pytorch_lightning[extra]==1.8.0.post1
timm==0.6.11
torch==1.13.0
torchvision==0.14.0
transformers==4.21.0
